Fixed an error I saw on windows machines, and an error encountered using dispersions.py with python 3.10
Former-commit-id: 81429e43a75b3ce7e1446c6b2f69cd4f05dec6ea
Former-commit-id: fb28669b62c0d0025011daac1148022af940479d

+++ src/fswAlgorithms/effectorInterfaces/rwMotorVoltage/_UnitTest/test_rwMotorVoltage.py
@@ -135,7 +135,7 @@ def run(show_plots, useLargeVoltage, useAvailability, useTorqueLoop, testName):
     # create RW availability message
     if useAvailability:
         rwAvailabilityMessage = messaging.RWAvailabilityMsgPayload()
-        rwAvailArray = np.zeros(messaging.MAX_EFF_CNT)
+        rwAvailArray = np.zeros(messaging.MAX_EFF_CNT, dtype=int)
         rwAvailArray.fill(messaging.AVAILABLE)
         rwAvailArray[2] = messaging.UNAVAILABLE        # make 3rd RW unavailable
         rwAvailabilityMessage.wheelAvailability = rwAvailArray
